+++ lib/src/main/java/com/github/xizzhu/simpletooltip/ToolTip.java
@@ -18,11 +18,12 @@ package com.github.xizzhu.simpletooltip;
 
 import android.graphics.Color;
 import android.graphics.Typeface;
-import android.support.annotation.ColorInt;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.annotation.StringRes;
 import android.view.Gravity;
+
+import androidx.annotation.ColorInt;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.annotation.StringRes;
 
 public class ToolTip {
     @StringRes